refactor: 상태에 따른 컴포넌트 분리
- quiz 상태(quiz title, content), visual 상태(docker visualization)에 따라 컴포넌트 분리
- quiz 상태는 QuizNodes에서 관리
- visual 상태는 VisualizationNodes에서 관리
